Backups/BluechipStock/requirements.txt
requests
pandas
beautifulsoup4
lxml
seaborn
matplotlib
pandas-ta
scikit-learn
tensorflow
nepali-datetime
joblib
Flask
APScheduler
Flask-CORS
Pillow
pytz
visualkeras